refactor: split pi-specdocs extension into modules
